+++ test/transformers/group_by_test.dart
@@ -10,13 +10,28 @@ void main() {
     await expectLater(
         Stream.fromIterable([1, 2, 3, 4]).groupBy((value) => value),
         emitsInOrder(<Matcher>[
-          TypeMatcher<GroupByStream<int, int>>()
+          TypeMatcher<GroupedStream<int, int>>()
               .having((stream) => stream.key, 'key', 1),
-          TypeMatcher<GroupByStream<int, int>>()
+          TypeMatcher<GroupedStream<int, int>>()
               .having((stream) => stream.key, 'key', 2),
-          TypeMatcher<GroupByStream<int, int>>()
+          TypeMatcher<GroupedStream<int, int>>()
               .having((stream) => stream.key, 'key', 3),
-          TypeMatcher<GroupByStream<int, int>>()
+          TypeMatcher<GroupedStream<int, int>>()
+              .having((stream) => stream.key, 'key', 4),
+          emitsDone
+        ]));
+
+    await expectLater(
+        Stream.fromIterable([1, 2, 3, 4])
+            .groupBy((value) => value, durationSelector: (_) => Rx.never()),
+        emitsInOrder(<Matcher>[
+          TypeMatcher<GroupedStream<int, int>>()
+              .having((stream) => stream.key, 'key', 1),
+          TypeMatcher<GroupedStream<int, int>>()
+              .having((stream) => stream.key, 'key', 2),
+          TypeMatcher<GroupedStream<int, int>>()
+              .having((stream) => stream.key, 'key', 3),
+          TypeMatcher<GroupedStream<int, int>>()
               .having((stream) => stream.key, 'key', 4),
           emitsDone
         ]));
@@ -26,6 +41,22 @@ void main() {
     await expectLater(
         Stream.fromIterable([1, 2, 3, 4])
             .groupBy((value) => _toEventOdd(value % 2))
+            .flatMap((stream) => stream.map((event) => {stream.key: event})),
+        emitsInOrder(<dynamic>[
+          {'odd': 1},
+          {'even': 2},
+          {'odd': 3},
+          {'even': 4},
+          emitsDone
+        ]));
+
+    await expectLater(
+        Stream.fromIterable([1, 2, 3, 4])
+            .groupBy(
+              (value) => _toEventOdd(value % 2),
+              durationSelector: (_) =>
+                  Stream.periodic(const Duration(seconds: 1)),
+            )
             .flatMap((stream) => stream.map((event) => {stream.key: event})),
         emitsInOrder(<dynamic>[
           {'odd': 1},
@@ -54,6 +85,28 @@ void main() {
           },
           emitsDone
         ]));
+
+    await expectLater(
+        Stream.fromIterable([1, 2, 3, 4])
+            .groupBy(
+              (value) => _toEventOdd(value % 2),
+              durationSelector: (_) =>
+                  Stream.periodic(const Duration(seconds: 1)),
+            )
+            // fold is called when onDone triggers on the Stream
+            .map((stream) async => await stream.fold(
+                {stream.key: <int>[]},
+                (Map<String, List<int>> previous, element) =>
+                    previous..[stream.key]?.add(element))),
+        emitsInOrder(<dynamic>[
+          {
+            'odd': [1, 3]
+          },
+          {
+            'even': [2, 4]
+          },
+          emitsDone
+        ]));
   });
 
   test('Rx.groupBy.emittedStreamCallOnDone', () async {
@@ -63,65 +116,189 @@ void main() {
             // drain will emit 'done' onDone
             .map((stream) async => await stream.drain('done')),
         emitsInOrder(<dynamic>['done', 'done', 'done', 'done', emitsDone]));
+
+    await expectLater(
+        Stream.fromIterable([1, 2, 3, 4])
+            .groupBy((value) => value, durationSelector: (_) => Rx.never())
+            // drain will emit 'done' onDone
+            .map((stream) async => await stream.drain('done')),
+        emitsInOrder(<dynamic>['done', 'done', 'done', 'done', emitsDone]));
   });
 
   test('Rx.groupBy.asBroadcastStream', () async {
-    final stream = Stream.fromIterable([1, 2, 3, 4])
-        .asBroadcastStream()
-        .groupBy((value) => value);
+    {
+      final stream = Stream.fromIterable([1, 2, 3, 4])
+          .asBroadcastStream()
+          .groupBy((value) => value);
 
-    // listen twice on same stream
-    stream.listen(null);
-    stream.listen(null);
-    // code should reach here
-    await expectLater(true, true);
+      // listen twice on same stream
+      stream.listen(null);
+      stream.listen(null);
+      // code should reach here
+      await expectLater(true, true);
+    }
+
+    {
+      final stream =
+          Stream.fromIterable([1, 2, 3, 4]).asBroadcastStream().groupBy(
+                (value) => value,
+                durationSelector: (_) =>
+                    Stream.periodic(const Duration(seconds: 2)),
+              );
+
+      // listen twice on same stream
+      stream.listen(null);
+      stream.listen(null);
+      // code should reach here
+      await expectLater(true, true);
+    }
   });
 
   test('Rx.groupBy.pause.resume', () async {
-    var count = 0;
-    late StreamSubscription subscription;
+    {
+      var count = 0;
+      late StreamSubscription subscription;
 
-    subscription = Stream.fromIterable([1, 2, 3, 4])
-        .groupBy((value) => value)
-        .listen(expectAsync1((result) {
-          count++;
+      subscription = Stream.fromIterable([1, 2, 3, 4])
+          .groupBy((value) => value)
+          .listen(expectAsync1((result) {
+            count++;
 
-          if (count == 4) {
-            subscription.cancel();
-          }
-        }, count: 4));
+            if (count == 4) {
+              subscription.cancel();
+            }
+          }, count: 4));
 
-    subscription.pause(Future<void>.delayed(const Duration(milliseconds: 100)));
+      subscription
+          .pause(Future<void>.delayed(const Duration(milliseconds: 100)));
+    }
+
+    {
+      var count = 0;
+      late StreamSubscription subscription;
+
+      subscription = Stream.fromIterable([1, 2, 3, 4])
+          .groupBy(
+            (value) => value,
+            durationSelector: (_) => Rx.timer(null, const Duration(seconds: 1)),
+          )
+          .listen(expectAsync1((result) {
+            count++;
+
+            if (count == 4) {
+              subscription.cancel();
+            }
+          }, count: 4));
+
+      subscription
+          .pause(Future<void>.delayed(const Duration(milliseconds: 100)));
+    }
   });
 
   test('Rx.groupBy.error.shouldThrow.onError', () async {
-    final streamWithError =
-        Stream<void>.error(Exception()).groupBy((value) => value);
+    {
+      final streamWithError =
+          Stream<void>.error(Exception()).groupBy((value) => value);
 
-    streamWithError.listen(null,
-        onError: expectAsync2((Exception e, StackTrace s) {
-      expect(e, isException);
-    }));
+      streamWithError.listen(null,
+          onError: expectAsync2((Exception e, StackTrace s) {
+        expect(e, isException);
+      }));
+    }
+
+    {
+      final streamWithError = Stream<void>.error(Exception()).groupBy(
+        (value) => value,
+        durationSelector: (_) => Rx.timer(null, const Duration(seconds: 1)),
+      );
+
+      streamWithError.listen(null,
+          onError: expectAsync2((Exception e, StackTrace s) {
+        expect(e, isException);
+      }));
+    }
   });
 
   test('Rx.groupBy.error.shouldThrow.onGrouper', () async {
-    final streamWithError = Stream.fromIterable([1, 2, 3, 4]).groupBy((value) {
-      throw Exception();
-    });
+    {
+      final streamWithError =
+          Stream.fromIterable([1, 2, 3, 4]).groupBy((value) {
+        throw Exception();
+      });
 
-    streamWithError.listen(null,
-        onError: expectAsync2((Exception e, StackTrace s) {
-          expect(e, isException);
-        }, count: 4));
+      streamWithError.listen(null,
+          onError: expectAsync2((Exception e, StackTrace s) {
+            expect(e, isException);
+          }, count: 4));
+    }
+
+    {
+      final streamWithError = Stream.fromIterable([1, 2, 3, 4]).groupBy(
+        (value) => throw Exception(),
+        durationSelector: (_) => Rx.timer(null, const Duration(seconds: 1)),
+      );
+
+      streamWithError.listen(null,
+          onError: expectAsync2((Exception e, StackTrace s) {
+            expect(e, isException);
+          }, count: 4));
+    }
   });
   test('Rx.groupBy accidental broadcast', () async {
-    final controller = StreamController<int>();
+    {
+      final controller = StreamController<int>();
 
-    final stream = controller.stream.groupBy((_) => _);
+      final stream = controller.stream.groupBy((_) => _);
 
-    stream.listen(null);
-    expect(() => stream.listen(null), throwsStateError);
+      stream.listen(null);
+      expect(() => stream.listen(null), throwsStateError);
 
-    controller.add(1);
+      controller.add(1);
+    }
+
+    {
+      final controller = StreamController<int>();
+
+      final stream = controller.stream.groupBy(
+        (_) => _,
+        durationSelector: (_) => Rx.timer(null, const Duration(seconds: 1)),
+      );
+
+      stream.listen(null);
+      expect(() => stream.listen(null), throwsStateError);
+
+      controller.add(1);
+    }
+  });
+
+  test('Rx.groupBy.durationSelector', () {
+    final g = [
+      '0 -> 1',
+      '1 -> 1',
+      '2 -> 1',
+      '0 -> 2',
+      '1 -> 2',
+      '2 -> 2',
+    ];
+    final take = 30;
+
+    final stream = Stream.periodic(const Duration(milliseconds: 100), (i) => i)
+        .groupBy(
+          (i) => i % 3,
+          durationSelector: (i) =>
+              Rx.timer(null, const Duration(milliseconds: 400)),
+        )
+        .flatMap((g) => g
+            .scan<int>((acc, value, index) => acc + 1, 0)
+            .map((event) => '${g.key} -> $event'))
+        .take(take);
+
+    expect(
+      stream,
+      emitsInOrder(<Object>[
+        ...List.filled(take ~/ g.length, g).expand<String>((e) => e),
+        emitsDone,
+      ]),
+    );
   });
 }