+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * Add optional parameter `maxConcurrent` to `flatMap`.
 * Update to docs ([@AlexanderJohr](https://github.com/AlexanderJohr)).
 * Bugfix: `onErrorReturnWith` drops remaining data event after the first error.
+* Update `groupBy`
+  * Rename `GroupByStream` to `GroupedStream`.
+  * Add optional parameter `durationSelector`, which used to determine how long each group should exist.
 
 ## 0.27.1